Premium Analytics: Inline Sales by UTM render component
