+++ projects/packages/premium-analytics/widgets/sales-by-utm-channel/helpers/build-sales-by-utm-data.ts
