+++ projects/packages/premium-analytics/widgets/sales-by-utm-channel/render.tsx
@@ -1,23 +1,79 @@
 /**
  * External dependencies
  */
-import { WidgetRoot } from '@jetpack-premium-analytics/widgets-toolkit';
-import type { ComponentProps } from 'react';
+import { useReportOrderAttribution } from '@jetpack-premium-analytics/data';
+import { channel } from '@jetpack-premium-analytics/icons';
+import {
+	LeaderboardChart,
+	WidgetLoadingOverlay,
+	WidgetRoot,
+	formatLegendLabels,
+	useWidgetError,
+	useWidgetRootContext,
+} from '@jetpack-premium-analytics/widgets-toolkit';
+import { useMemo, type ComponentProps, type CSSProperties } from 'react';
 /**
  * Internal dependencies
  */
-import { SalesByUtmWidget } from './sales-by-utm-widget';
+import { buildSalesByUtmData } from './helpers/build-sales-by-utm-data';
 
 type SalesByUtmChannelRenderProps = Pick< ComponentProps< typeof WidgetRoot >, 'attributes' > & {
 	setError?: ComponentProps< typeof WidgetRoot >[ 'setError' ];
 };
 
+function SalesByUtmChannelWidget() {
+	const { reportParams } = useWidgetRootContext();
+
+	const params = useMemo(
+		() => ( {
+			...reportParams,
+			view: 'channel' as const,
+		} ),
+		[ reportParams ]
+	);
+
+	const { primary, hasComparison, isLoading, isFetching, hasData, isError, error, refetch } =
+		useReportOrderAttribution( params );
+
+	const isInitialLoading = isLoading && ! hasData;
+	const isRefetching = isFetching && hasData;
+
+	const chartData = useMemo( () => buildSalesByUtmData( primary.data ), [ primary.data ] );
+	const legendLabels = useMemo( () => formatLegendLabels( reportParams ), [ reportParams ] );
+
+	const hasError = useWidgetError( isError, error, refetch );
+	if ( hasError ) {
+		return null;
+	}
+
+	if ( isInitialLoading ) {
+		return <WidgetLoadingOverlay />;
+	}
+
+	return (
+		<>
+			<LeaderboardChart
+				data={ chartData }
+				withComparison={ hasComparison }
+				legendLabels={ legendLabels }
+				emptyStateIcon={ channel }
+				style={
+					{
+						'--a8c--charts--leaderboard--bar--border-radius': '0 1px 1px 0',
+					} as CSSProperties
+				}
+			/>
+			{ isRefetching && <WidgetLoadingOverlay /> }
+		</>
+	);
+}
+
 /**
  * Sales by UTM channel widget.
  *
  * WidgetRoot provides the query client, chart theme, and resolved report params;
- * SalesByUtmWidget composes toolkit primitives to fetch the order-attribution
- * report and render the channel leaderboard.
+ * this render module fetches the order-attribution report and renders the
+ * channel leaderboard.
  *
  * @param root0            - Component props.
  * @param root0.attributes - Widget attributes.
@@ -30,7 +86,7 @@ export default function SalesByUtmChannelRender( {
 }: SalesByUtmChannelRenderProps ) {
 	return (
 		<WidgetRoot attributes={ attributes } setError={ setError } options={ { from: '/' } }>
-			<SalesByUtmWidget view="channel" />
+			<SalesByUtmChannelWidget />
 		</WidgetRoot>
 	);
 }
